third_party/blink/web_tests/platform/linux-chrome/external/wpt/web-bundle/subresource-loading/csp-blocked.https.tentative-expected.txt

This is a testharness.js-based test.
[FAIL] Check the CSP violation reports.
  assert_equals: Report count. expected 2 but got 0
Harness: the test ran to completion.
